src/skill/requirements.txt
python-dotenv
botbuilder-integration-aiohttp>=4.15.0
teams-ai>=1.4.0,<2.0.0
semantic-kernel>=1.22.0
dapr>=1.14.0,<2.0.0
